Fix group comparison
The comparison logic for ApiGroup was not working. This has led to the
incorrect detemination of "current version" versus "old versions" for a
Definition. This PR fixes it.
