+++ gen-apidocs/generators/config.yaml
@@ -142,7 +142,7 @@ resource_categories:
       description_warning: "Alpha objects should not be used in production and may not be compatible with future versions of the resource type."
     - name: PodSecurityPolicy
       version: v1beta1
-      group: extensions
+      group: policy
   - name: "Cluster APIs"
     include: "cluster"
     resources:
